+++ back/app/src/main/java/org/example/repository/CenterRepository.java
@@ -8,10 +8,10 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface CenterRepository extends JpaRepository<Center, Integer> {
+public interface CenterRepository extends JpaRepository<Center, Long> {
     @Query("SELECT c FROM Center c WHERE LOWER(c.city) LIKE LOWER(CONCAT('%', :city, '%'))")
     List<Center> findByCityLike(@Param("city") String city);
 
     @Query("SELECT s FROM Specialist s WHERE s.center.id = :centerId")
-    List<Specialist> findSpecialistsByCenterId(@Param("centerId") Integer centerId);
+    List<Specialist> findSpecialistsByCenterId(@Param("centerId") Long centerId);
 }
